9.inner class to upper level.
